fix volume increase/decrease value; fix stream load not pausing
Fixes a bug that stored a volume value as a string in localStorage;
this prevented a valid float value from being created when adding
a saved volume value (of type string) with an increment or decrement
value.

Also fixes a bug that attempted to pause a stream when a new stream
was loaded, after the new stream's data had been set; resulting in
the old stream not being actually paused by the Pause() method.